Refactors delegates functions, the querys are on a querys file now and the utils function on a delegates file Adds test file for delegate utils
