+++ server/delegate/delegate.route.js
@@ -5,13 +5,13 @@ const paramValidation = require('../../config/param-validation')
 const router = express.Router() // eslint-disable-line new-cap
 const { getByAddress, getROI } = require('./delegate.controller')
 
-/** GET /api/delegate/address/:address - Get delegate by address */
+/** GET /api/delegates/address/:address - Get delegate by address */
 router
-  .route('/:address')
+  .route('/address/:address')
 
   .get(validate(paramValidation.getByAddress), getByAddress)
 
-/** GET /api/delegate/roi/:address - Get delegate ROI by address */
+/** GET /api/delegates/roi/:address - Get delegate ROI by address */
 router
   .route('/roi/:address')
